use_case_1/requirements.txt
python-dotenv
openai
python-docx
pandas
python-dotenv
openpyxl
ruff